Added tilt and whammy support
With the tilt support added, the method I chose for forcing clone hero to assign the virtual controller and not the wiimote works. The methods for calibrating the tilt and whammy controls act similar to wiitarthing where you press 1 on the wiimote to set the min tilt, and 2 to set the max tilt. You also just move the whammy bar across its entire range of motion to calibrate it like you do with wiitarthing. Feel free to edit the default values for both the whammy and tilt controls, as I didn't really do much testing and the user will most likely calibrate beforehand and not use defaults anyway.